bismillah fix

+++ website/views.py
@@ -1,36 +1,82 @@
 from flask import Blueprint, render_template, request, redirect, url_for
 from flask_login import login_user, logout_user, current_user
-from . import db
-from .models import Admin, User, Kerusakan, MyQueue
+from . import db, app
+from .models import Admin, User, Kerusakan, Nota
 import sqlite3
+import os
 
 views = Blueprint('views', __name__)
 
+picFolder = os.path.join('static', 'pics')
+app.config['UPLOAD_FOLDER'] = picFolder
+
 @views.route('/', methods=['GET', 'POST'])
 def user():
-    q = MyQueue()
-    antrian = q.enqueue()
-    pengambilan = q.dequeue()
+    pic1= os.path.join(app.config['UPLOAD_FOLDER'], '1.png')
+    pic2 = os.path.join(app.config['UPLOAD_FOLDER'], '2.png')
+    pic3 = os.path.join(app.config['UPLOAD_FOLDER'], 'montir.png')
 
-    if request.method == "POST":
-        data = request.form.get("antrian")
-        kerusakan = request.form.get('kerusakan')
-
-        new_user = User(antrian=data)
+    conn = sqlite3.connect("./instance/db_antrian.db")
+    curs = conn.cursor()
+    curs.execute('SELECT * FROM user')
+    cek = curs.fetchall()
+    if cek == []:
+        new_user = User(antrian=0)
         db.session.add(new_user)
         db.session.commit()
-
-        new_kerusakan = Kerusakan(kerusakan=kerusakan, user_antrian=data)
-        db.session.add(new_kerusakan)
-        db.session.commit()
-
+        curs.close()
         return redirect(url_for('views.user'))
+    else:
+        cek_head = curs.execute('SELECT * FROM user ORDER by antrian asc')
+        head = cek_head.fetchone()[1]
 
-    return render_template("home.html",antrian=antrian, pengambilan=pengambilan, user=current_user)
+        cek_tail = curs.execute('SELECT * FROM user ORDER by antrian desc')
+        tail = cek_tail.fetchone()[1]
+
+        cek_nota = curs.execute('SELECT * FROM nota WHERE user_antrian="{}"'.format(head))
+        nota = cek_nota.fetchmany()
+        curs.close()
+        if nota == []:
+            nota_nama = '-'
+            nota_motor = '-'
+            nota_plat = '-'
+            nota_kerusakan = '-'
+            nota_biaya = '-'
+        else:
+            for row in nota:
+                nota_nama = row[1]
+                nota_motor = row[2]
+                nota_plat = row[3]
+                nota_kerusakan = row[4]
+                nota_biaya = row[5]
+
+        if request.method == "POST":
+            data = request.form.get("antrian")
+            nama = request.form.get('nama')
+            motor = request.form.get('motor')
+            plat = request.form.get('plat')
+
+            list_kerusakan = request.form.getlist('servicecbx')
+            kerusakan = ''
+            for rusak in list_kerusakan:
+                kerusakan += rusak + ', '
+
+            new_user = User(antrian=data)
+            new_kerusakan = Kerusakan(nama=nama, motor=motor, plat=plat, kerusakan=kerusakan, user_antrian=data)
+            db.session.add(new_user)
+            db.session.add(new_kerusakan)
+            db.session.commit()
+            return redirect(url_for('views.user'))
+
+        return render_template("home.html", img1=pic1, img2=pic2, img3=pic3, antrian=tail+1, panggil=head, user=current_user, nota_nama=nota_nama, nota_motor=nota_motor, nota_plat=nota_plat,nota_kerusakan=nota_kerusakan, nota_biaya=nota_biaya)
 
 
 @views.route('/login', methods=['GET', 'POST'])
 def login():
+    pic1= os.path.join(app.config['UPLOAD_FOLDER'], '1.png')
+    pic2 = os.path.join(app.config['UPLOAD_FOLDER'], '2.png')
+    pic3 = os.path.join(app.config['UPLOAD_FOLDER'], 'montir.png')
+
     email = request.form.get('email')
     password = request.form.get('password')
 
@@ -40,27 +86,74 @@ def login():
             login_user(user)
             return redirect(url_for('views.admin'))
 
-    return render_template('login.html', user=current_user)
+    return render_template('login.html', user=current_user, img1=pic1, img2=pic2, img3=pic3)
 
-@views.route('/admin', methods=['GET', 'POST'])
+
+@views.route('/admin', methods=['GET', 'POST', 'DELETE'])
 def admin():
-    q = MyQueue()
-    antrian = q.enqueue()
-    panggil = q.dequeue()
-    # rusak = Kerusakan.query.filter_by(user_antrian=5)
+    pic1= os.path.join(app.config['UPLOAD_FOLDER'], '1.png')
+    pic2 = os.path.join(app.config['UPLOAD_FOLDER'], '2.png')
+    pic3 = os.path.join(app.config['UPLOAD_FOLDER'], 'montir.png')
+
     conn = sqlite3.connect("./instance/db_antrian.db")
     curs = conn.cursor()
-    # kerusakan = curs.execute('SELECT kerusakan FROM kerusakan WHERE user_antrian = 5')
-    
-    if request.method=='POST':
-        data = request.form.get('panggil')
-        User.query.filter_by(antrian=data).delete()
-        Kerusakan.query.filter_by(user_antrian=data).delete()
+    curs.execute('SELECT * FROM user')
+    cek = curs.fetchall()
+    if cek == []:
+        new_user = User(antrian=0)
+        db.session.add(new_user)
         db.session.commit()
-        
+        curs.close()
         return redirect(url_for('views.admin'))
 
-    return render_template('admin.html', user=current_user, panggil=panggil, antrian=antrian)  
+    curs.execute('SELECT * FROM user ORDER by antrian asc')
+    head = curs.fetchone()[1]
+
+    cek = curs.execute('SELECT * FROM kerusakan WHERE user_antrian="{}"'.format(head))
+    cek = cek.fetchall()
+    if cek == []:
+        new_kerusakan = Kerusakan(nama='', motor='', plat='', kerusakan='', user_antrian=0)
+        db.session.add(new_kerusakan)
+        db.session.commit()
+        curs.close()
+        return redirect(url_for('views.admin'))
+
+
+    cek_next = curs.execute('SELECT * FROM kerusakan WHERE user_antrian="{}"'.format(head+1))
+    cek_next = cek_next.fetchall()
+    if cek_next == []:
+        nama = '-'
+        motor = '-'
+        plat = '-'
+        kerusakan = '-'
+    else:
+        for row in cek_next:
+            nama = row[1]
+            motor = row[2]
+            plat = row[3]
+            kerusakan = row[4]
+    if request.method=='POST':
+        if request.form.get('next') == 'PANGGIL':
+            data = request.form.get('panggil')
+            total = request.form.get('total')
+            
+            User.query.filter_by(antrian=data).delete()
+            Kerusakan.query.filter_by(user_antrian=data).delete()
+            new_nota = Nota(nama=nama, motor=motor, plat=plat, kerusakan=kerusakan, biaya=int(total), user_antrian=int(data)+1)
+            db.session.add(new_nota)
+            db.session.commit()
+
+            return redirect(url_for('views.admin'))
+
+        if request.form.get('delete') == 'RESET NOTA':
+            cek = Nota.query.all()
+            for i in range(len(cek)+1):
+                Nota.query.filter_by(id=(i+1)).delete()
+                db.session.commit()
+            return redirect(url_for('views.admin'))
+
+    return render_template('admin.html', img1=pic1, img2=pic2, img3=pic3, user=current_user, panggil=head, nama=nama, motor=motor, plat=plat, kerusakan=kerusakan)
+
 
 @views.route('/logout')
 def logout():
